test: move feature gating tests to each module's __tests__/ directory
Replace centralized test file with per-module feature-gate.test.ts files
alongside each business module, following the project's test conventions.

Tests: 16 pass across 7 files (absences, accidents, warnings,
medical-certificates, terminations, ppe-deliveries, employees).
